Rename simpleKeyStore composable and update imports
