+++ client/src/composables/simpleKeyStore.ts
