+++ client/src/stores/collectionAttributesStore.ts
@@ -2,7 +2,7 @@ import { defineStore } from "pinia";
 
 import type { DatasetCollectionAttributes } from "@/api";
 import { fetchCollectionAttributes } from "@/api/datasetCollections";
-import { useSimpleKeyStore } from "@/composables/simpleStore";
+import { useSimpleKeyStore } from "@/composables/simpleKeyStore";
 
 export const useCollectionAttributesStore = defineStore("collectionAttributesStore", () => {
     const { storedItems, getItemById, isLoadingItem } = useSimpleKeyStore<DatasetCollectionAttributes>((params) =>
