+++ client/src/stores/datasetExtraFilesStore.ts
@@ -1,7 +1,7 @@
 import { defineStore } from "pinia";
 
 import { DatasetExtraFiles, fetchDatasetExtraFiles } from "@/api/datasets";
-import { useSimpleKeyStore } from "@/composables/simpleStore";
+import { useSimpleKeyStore } from "@/composables/simpleKeyStore";
 
 export const useDatasetExtraFilesStore = defineStore("datasetExtraFilesStore", () => {
     const { storedItems, getItemById, isLoadingItem, fetchItemById } = useSimpleKeyStore<DatasetExtraFiles>((params) =>
